requirements.txt
google-adk>=0.1.0
google-cloud-speech>=2.21.0
google-cloud-texttospeech>=2.14.0
google-cloud-translate>=3.12.0
google-genai>=0.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
websockets>=12.0
numpy>=1.24.0